requirements.txt
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
replicate>=0.25.0
pyyaml>=6.0

